debrepo-install: add ability to add repo without install

+++ travis/build/debrepo-install.sh
@@ -7,9 +7,13 @@ set -x
 # Install .deb packages from release branch
 echo deb [trusted=yes] https://cafe-desktop.github.io/debian-packages/ ./ >> /etc/apt/sources.list
 aptitude -y update
-aptitude -y install $1 > aptlog
-if grep -w "unmet dependencies" aptlog; then
-  cat aptlog
-  exit 1
+
+if [ $# -gt 0 ]
+then
+  aptitude -y install $1 > aptlog
+  if grep -w "unmet dependencies" aptlog; then
+    cat aptlog
+    exit 1
+  fi
+  rm aptlog
 fi
-rm aptlog
